% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_hk_tide.R
\name{dict_hk_tide}
\alias{dict_hk_tide}
\title{Dictionary of HK tidal stations}
\usage{
dict_hk_tide()
}
\value{
\code{tibble}, with columns \code{name}, \code{full_name}, \code{lon}, \code{lat}, \code{available}
}
\description{
Provides the dictionary of HK tidal stations, with the following columns
\itemize{
\item name = Code name of the tidal stations
\item full_name = Name of the tidal stations, in English
\item lon = Longitude of the tidal stations
\item lat = Latitude of the tidal stations
\item available = Availability of the tidal stations, \code{FALSE} represents the tidal station is not operational, \code{TRUE} represents the tidal station is operational and will provide real-time tidal data to HKO
}
}
\examples{
dict_hk_tide()
}
