% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_hk_wxph.R
\name{dict_hk_wxph}
\alias{dict_hk_wxph}
\title{Dictionary of HK tidal stations}
\usage{
dict_hk_wxph()
}
\value{
\code{tibble}, with columns \code{code}, \code{full_name}
}
\description{
Provides the dictionary of HKO weather camera stations, with the following columns
\itemize{
\item code = Code name of the tidal stations
\item full_name = Name of the tidal stations, in English
}
}
\examples{
dict_hk_wxph()
}
