% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_time.R
\name{hk_time}
\alias{hk_time}
\title{Call China/ Hong Kong/ Macau/ Taiwan Standard Time}
\usage{
hk_time()
}
\value{
POSIXct, in HKT
}
\description{
Hong Kong Standard Time (HKT) will be called by this function
}
\examples{
hk_time()
}
