% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cn_radr2.R
\name{load_cn_radr_official}
\alias{load_cn_radr_official}
\title{CN weather - Downloading RADAR data (official)}
\usage{
load_cn_radr_official(
  time = seq.POSIXt(from = wx3hk::hk_time(), tz = "HongKong", by = "-6 mins", length.out
    = 1680),
  zoom = "*all*",
  list_fail = TRUE,
  dir = getwd(),
  attempt = 2L,
  worker = 1L
)
}
\arguments{
\item{time}{(POSIXt) An array of date-time to be downloaded.}

\item{zoom}{(character) Radar photo zoom range, the following values are accepted
\itemize{
\item \emph{all}: All of the following zoom ranges will be downloaded
\item CHN: Entire China
\item NCN: Northern China
\item NEC: North-Eastern China
\item ECN: Eastern China
\item CCN: Central China
\item SCN: Southern China
\item SWC: South-Western China
\item NWC: North-Western China
}}

\item{list_fail}{(logical) List failed-to-download items.
\itemize{
\item TRUE for listing all failed-to-download items. (DEFAULT)
\item FALSE for not listing all failed-to-download items.
}}

\item{dir}{(character) Directory of downloaded data. A folder (HK_Data) will be created in the directory provided to store the aforementioned data.}

\item{attempt}{(integer) Attempts to be made per file to download.}

\item{worker}{(integer) Number of sessions to be open to download a list of files. Files will be downloaded simultaneously.}
}
\value{
Images from http://www.nmc.cn/
}
\description{
RADAR images of the entire China will be provided. Images are downloaded from http://www.nmc.cn/.
}
\examples{
load_cn_radr()
}
