% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hk_all.R
\name{load_hk_all}
\alias{load_hk_all}
\title{Download (all) available HK data}
\usage{
load_hk_all(worker = 1L, list_fail = TRUE)
}
\arguments{
\item{worker}{(integer) Number of sessions to be open to download a list of files. Files will be downloaded simultaneously.}

\item{list_fail}{(logical) List failed-to-download items.
\itemize{
\item TRUE for listing all failed-to-download items. (DEFAULT)
\item FALSE for not listing all failed-to-download items.
}}
}
\value{
Data downloaded from most functions under \code{?wx3hk::load_hk}
}
\description{
This function downloads all of the available HK data present in this package \code{wx3hk}. To understand what data is available, please check \code{?wx3hk::load_hk}
}
\examples{
wx3hk::load_hk_all()
}
