% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hk_mslp.R
\name{load_hk_mslp}
\alias{load_hk_mslp}
\title{HK weather - Downloading mean sea level pressure data}
\usage{
load_hk_mslp(
  time = seq.POSIXt(from = Sys.time(), by = "-10 min", length.out = 1009),
  lan = "en",
  list_fail = TRUE,
  dir = getwd(),
  attempt = 5L,
  worker = 1L
)
}
\arguments{
\item{time}{(POSIXt) An array of date-time to be downloaded.}

\item{lan}{(character) Language to download. Accepts the following
\itemize{
\item "en" for English
\item "tc" for Traditional Chinese
\item "sc" for Simplified Chinese
}}

\item{list_fail}{(logical) List failed-to-download items.
\itemize{
\item TRUE for listing all failed-to-download items. (DEFAULT)
\item FALSE for not listing all failed-to-download items.
}}

\item{dir}{(character) Directory of downloaded data. A folder (\code{HK_Data}) will be created in the directory provided to store the aforementioned data.}

\item{attempt}{(integer) Attempts to be made per file to download.}

\item{worker}{(integer) Number of sessions to be open to download a list of files. Files will be downloaded simultaneously.}
}
\value{
CSVs from https://data.gov.hk/en/help/api-spec
}
\description{
Downloading mean sea level pressure data from https://data.gov.hk/en/help/api-spec
}
\examples{
load_hk_mslp()
}
