% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hk_rainHr.R
\name{load_hk_rainHr}
\alias{load_hk_rainHr}
\title{HK weather - Downloading daily rainfall distribution images}
\usage{
load_hk_rainHr(
  time = seq.POSIXt(from = wx3hk::hk_time(), by = "-15 min", length.out = 672),
  lan = "en",
  list_fail = TRUE,
  dir = getwd(),
  attempt = 5L,
  worker = 1L
)
}
\arguments{
\item{time}{(POSIXt) An array of date-time to be downloaded.}

\item{lan}{(character) Language to download. Accepts the following
\itemize{
\item "en" for English
\item "tc" for Traditional Chinese
\item "sc" for Simplified Chinese
}}

\item{list_fail}{(logical) List failed-to-download items.
\itemize{
\item TRUE for listing all failed-to-download items. (DEFAULT)
\item FALSE for not listing all failed-to-download items.
}}

\item{dir}{(character) Directory of downloaded data. A folder (\code{HK_Data}) will be created in the directory provided to store the aforementioned data.}

\item{attempt}{(integer) Attempts to be made per file to download.}

\item{worker}{(integer) Number of sessions to be open to download a list of files. Files will be downloaded simultaneously.}
}
\value{
Images from https://www.hko.gov.hk/en/index.html
}
\description{
Download data from https://www.hko.gov.hk/en/index.html
}
\examples{
load_hk_rainHr()
}
