% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hk_tide.R
\name{load_hk_tide}
\alias{load_hk_tide}
\title{HK weather - Downloading tidal height data}
\usage{
load_hk_tide(
  time = seq.POSIXt(from = wx3hk::hk_time(), by = "-5 min", length.out = 2016),
  lan = "en",
  type = c("hko", "md"),
  list_fail = TRUE,
  dir = getwd(),
  attempt = 5L,
  worker = 1L
)
}
\arguments{
\item{time}{(POSIXt) An array of date-time to be downloaded.}

\item{lan}{(character) Language to download. Accepts the following
\itemize{
\item "en" for English
\item "tc" for Traditional Chinese
\item "sc" for Simplified Chinese
}}

\item{type}{(character) Source of data to be downloaded. Accepts the following (any combination)
\itemize{
\item "hko" for Hong Kong Observatory
\item "md" for Marine Department
}}

\item{list_fail}{(logical) List failed-to-download items.
\itemize{
\item TRUE for listing all failed-to-download items. (DEFAULT)
\item FALSE for not listing all failed-to-download items.
}}

\item{dir}{(character) Directory of downloaded data. A folder (\code{HK_Data}) will be created in the directory provided to store the aforementioned data.}

\item{attempt}{(integer) Attempts to be made per file to download.}

\item{worker}{(integer) Number of sessions to be open to download a list of files. Files will be downloaded simultaneously.}
}
\value{
CSVs from https://data.gov.hk/en/help/api-spec
}
\description{
Downloaded from https://data.gov.hk/en/help/api-spec
}
\examples{
load_hk_tide()
}
