% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hk_tidePredict.R
\name{load_hk_tidePredict}
\alias{load_hk_tidePredict}
\title{HK weather - Download predicted tidal height data}
\usage{
load_hk_tidePredict(station, year = 2015L:2025L)
}
\arguments{
\item{station}{(character) Array of station name abbreviations to be downloaded. Check \code{wx3hk::dict_hk_tide()} for more information}

\item{year}{(Integer) Array of years to download. Accepts value larger than 2015L}
}
\value{
\code{tibble} with columns \code{time}, \code{station}, \code{tide}
}
\description{
Downloaded from https://www.hko.gov.hk/en/tide/ttext.htm
}
\examples{
load_hk_tidePredict(station = "CLK", c(2016L, 2021L))
}
