% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hk_tidePredict_engine.R
\name{load_hk_tidePredict_2015HR}
\alias{load_hk_tidePredict_2015HR}
\title{Download hourly predicted tidal height data, using the year 2015 method}
\usage{
load_hk_tidePredict_2015HR(station, year)
}
\arguments{
\item{station}{Station to download. Check the \code{dict_hk_tide} for more information}

\item{year}{Year to download}
}
\description{
Download hourly predicted tidal height data, using the year 2015 method
}
\examples{
N/A
}
\keyword{internal}
