% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hk_tideSnapshot.R
\name{load_hk_tideSnap}
\alias{load_hk_tideSnap}
\title{HK weather - Downloading tidal height data (snapshots)}
\usage{
load_hk_tideSnap(
  station = dplyr::filter(wx3hk::dict_hk_tide(), available)$name,
  list_fail = TRUE,
  dir = getwd(),
  attempt = 5L,
  worker = 1L
)
}
\arguments{
\item{station}{(character) Array of station name abbreviations to be downloaded. Check \code{wx3hk::dict_hk_tide()} for more information}

\item{list_fail}{(logical) List failed-to-download items.
\itemize{
\item TRUE for listing all failed-to-download items. (DEFAULT)
\item FALSE for not listing all failed-to-download items.
}}

\item{dir}{(character) Directory of downloaded data. A folder (\code{HK_Data}) will be created in the directory provided to store the aforementioned data.}

\item{attempt}{(integer) Attempts to be made per file to download.}

\item{worker}{(integer) Number of sessions to be open to download a list of files. Files will be downloaded simultaneously.}
}
\value{
CSV files from https://www.hko.gov.hk/en/tide/marine/realtide.htm. Each CSV file corresponding to a snapshot of one of the stations
}
\description{
Downloaded from https://www.hko.gov.hk/en/tide/marine/realtide.htm

Snapshots of tidal height data provides more stations than normal API method. However, the tidal data must be snapshot once per day and lacks quality-control.
}
\examples{
load_hk_tideSnap()
}
