% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hk_wcht.R
\name{load_hk_wcht}
\alias{load_hk_wcht}
\title{HK weather - Downloading weather chart}
\usage{
load_hk_wcht(
  time = seq.POSIXt(from = Sys.time(), by = "-1 hour", length.out = 168),
  list_fail = TRUE,
  dir = getwd(),
  attempt = 5L,
  worker = 1L
)
}
\arguments{
\item{time}{(POSIXt) An array of date-time to be downloaded.}

\item{list_fail}{(logical) List failed-to-download items.
\itemize{
\item TRUE for listing all failed-to-download items. (DEFAULT)
\item FALSE for not listing all failed-to-download items.
}}

\item{dir}{(character) Directory of downloaded data. A folder (\code{HK_Data}) will be created in the directory provided to store the aforementioned data.}

\item{attempt}{(integer) Attempts to be made per file to download.}

\item{worker}{(integer) Number of sessions to be open to download a list of files. Files will be downloaded simultaneously.}
}
\value{
Images of Hong Kong weather chart, from https://www.hko.gov.hk/en/index.html
}
\description{
Downloaded from https://www.hko.gov.hk/en/index.html
}
\examples{
load_hk_wcht()
}
