% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hk_wxph.R
\name{load_hk_wxph}
\alias{load_hk_wxph}
\title{HK weather - Downloading weather photo}
\usage{
load_hk_wxph(
  time = seq.POSIXt(from = Sys.time(), by = "-5 min", length.out = 432),
  station = "all",
  list_fail = TRUE,
  dir = getwd(),
  attempt = 5L,
  worker = 1L
)
}
\arguments{
\item{time}{(POSIXt) An array of date-time to be downloaded.}

\item{station}{(character) An array of stations selected for task or \emph{special value} \code{all} to download all stations. Read \code{wx3hk::dict_hk_wxph()} for more information}

\item{list_fail}{(logical) List failed-to-download items.
\itemize{
\item TRUE for listing all failed-to-download items. (DEFAULT)
\item FALSE for not listing all failed-to-download items.
}}

\item{dir}{(character) Directory of downloaded data. A folder (\code{HK_Data}) will be created in the directory provided to store the aforementioned data.}

\item{attempt}{(integer) Attempts to be made per file to download.}

\item{worker}{(integer) Number of sessions to be open to download a list of files. Files will be downloaded simultaneously.}
}
\value{
Images from https://www.hko.gov.hk/en/index.html
}
\description{
Download weather photo from https://www.hko.gov.hk/en/index.html
}
\examples{
load_hk_wxph()
}
