% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mo_tide.R
\name{load_mo_tide}
\alias{load_mo_tide}
\title{MO weather - Download tidal images}
\usage{
load_mo_tide(
  time = seq.POSIXt(from = wx3hk::hk_hour(), tz = "Asia/Macau", by = "-1 hour",
    length.out = 91),
  dir = getwd(),
  attempt = 900L,
  subattempt = 5L,
  worker = 1L,
  list_fail = TRUE
)
}
\arguments{
\item{time}{(POSIXt) An array of date-time to be downloaded.}

\item{dir}{(character) Directory of downloaded data. A folder (\code{HK_Data}) will be created in the directory provided to store the aforementioned data.}

\item{attempt}{(integer) Attempts to be made per file to download.}

\item{subattempt}{(integer) Attempts to be made per attempt to download.}

\item{worker}{(integer) Number of sessions to be open to download a list of files. Files will be downloaded simultaneously.}

\item{list_fail}{(logical) List failed-to-download items.
\itemize{
\item TRUE for listing all failed-to-download items. (DEFAULT)
\item FALSE for not listing all failed-to-download items.
}}
}
\value{
Images from https://www.smg.gov.mo/zh
}
\description{
Data downloaded from https://www.smg.gov.mo/zh
}
\examples{
load_mo_tide()
}
