% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckf_HKLoad.R
\name{sys_ckf_HKLoadLan}
\alias{sys_ckf_HKLoadLan}
\title{HK/MOLoad language standard check}
\usage{
sys_ckf_HKLoadLan(lan)
}
\arguments{
\item{lan}{lan within the function}
}
\value{
NA
}
\description{
Check if language is supported
}
\examples{
sys_ckf_HKLoadLan(lan)
}
\keyword{internal}
